More Changes to V2 Architecture
Lists of damage and heals events are now properties on a player's PlayerCard. This avoids giant lists of every event for more than one player at a time and is more OOP in nature. Caching Architecture maintains a list of entities discovered to be actual players base on what they cast.